Copy configuration resources into container vs mounting
This makes configuration management more portable and helps on the Mac
where the docker daemon gets confused with /tmp, /private, and /var being
configured as places where files can be mounted from. The symlinks on the
Mac in those directories just seem to totally confuse Docker on the Mac.
This alleviates that problem.