Allow the shaded artifact to be the primary artifact
This will result in a testcontainers-concord artifact the is ~20mb, but
allows the embedding in a client environment that may share many of
the same dependencies.

This more closely resembles how you would use concord in production
where it is a separate system that has little chance of conflicting
with your environment.